Changed calculation algorithm & Fixed some problems

+++ Source/Othello/OthelloBlockGrid.h
@@ -97,7 +97,7 @@ public:
 	int GetRoundNumber();
 
 	UFUNCTION(BlueprintPure, Category = "Tensorflow")
-	TArray<int> GetState();
+	TArray<float> GetState();
 
 	UFUNCTION(BlueprintPure, Category = "Tensorflow")
 	bool IsGameOver();
